requirements.txt
web3
python-dotenv
langchain-openai
colorama